backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
google-genai>=0.3.0
werkzeug>=3.0.0
gunicorn>=21.2.0
